requirements.txt
google-generativeai 
numpy 
python-dotenv 
streamlit 
langchain
scikit-learn
gTTS